test(patina): split warn custom block coverage
